v1.5.0
option: -ts

Add last modified date timestamp info to a filename. This is done in a
shown in the sortable format of “Year-Month-Day Hour:Minute:Second”
thus if you choose to add this to the beginning of a filename, then
when you sort  a folder by name it will be shown in order.

-ts will prepend the timestamp by default, but you can use the -sa to
place it at the end of a filename (but before the file extension).