1.5.8 timestamp is working again
Due to a typo added long ago, this feature stopped working that I just
noticed. Oddly, what I was testing was looking for a bug dealing with
-sp -sa and -sn on a OS X system where those features were not working,
yet I am un-able to duplicate those problems on my own macs. With that
said, on that mac having issues running a dirty script looping printing
the results if I told it to count from 1 to 1000, it would skip groups
of numbers, so maybe perl is corrupt on that system or that system has
some sort of corruption itself.